requirements.txt
asgiref==3.5.2
coverage==6.4.1
Django==4.0.4
django-debug-toolbar==3.4.0
django-phonenumber-field==6.1.0
phonenumbers==8.12.49
Pillow==9.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
sqlparse==0.4.2
